+++ web/packages/dmt-core/src/components/index.ts
@@ -1,7 +1,7 @@
 export * from './Button'
 export * from './JsonView'
 export * from './Pickers'
-export * from './AccessControlList'
+export * from './AccessControl/AccessControlList'
 export * from './Select'
 export * from './UiPluginSelector'
 export * from './UploadFileButton'
