+++ web/packages/dmt-core/src/services/api/configs/gen/.openapi-generator/FILES
@@ -1,6 +1,5 @@
 .gitignore
 .npmignore
-.openapi-generator-ignore
 api.ts
 api/access-control-api.ts
 api/blob-api.ts
