+++ web/packages/dmt-core/src/services/api/configs/gen/models/access-level.ts
@@ -20,9 +20,9 @@
  * @enum {string}
  */
 export enum AccessLevel {
-    WRITE = 2,
-    READ = 1,
-    NONE = 0
+    WRITE = 'WRITE',
+    READ = 'READ',
+    NONE = 'NONE'
 }
